Partly converted GPU stuff - doesn't work yet

+++ src/BeatCalculatorParallel/BeatCalculatorParallel.h
@@ -15,7 +15,7 @@ class BeatCalculatorParallel {
         int filterbank();
         int diffrect();
         void cleanup(mpg123_handle* mh);
-        int readMP3(char* song, unsigned short* sample);
+        int readMP3(char* song, float* sample);
         void fftrArray(unsigned short* sample, int size, kiss_fft_cpx* out);
         void fftArray(unsigned short* sample, int size, kiss_fft_cpx* out);
         int combfilter(kiss_fft_cpx* fft_array, int size, int sample_size);
